refactor: simplify tunnel lifecycle and contracts

+++ src/main/java/com/huawei/devbridge/relaycontroller/domain/model/Tunnel.java
@@ -15,7 +15,7 @@ public class Tunnel {
     private String tunnelId;
     private Long tunnelCode;
     private String clusterId;
-    private Integer expiration;
+    private Long expiration;
     private String namespace;
     private String description;
     private Long bandwidthUsed;
